chore: move ErrorBoundary / QRCode / locales / useSnackbarCallback to shared from them
