+++ packages/dashboard/package.json
@@ -11,7 +11,7 @@
     "build": "build -- snowpack build"
   },
   "dependencies": {
-    "@dimensiondev/holoflows-kit": "0.8.0-20210317064617-6c4792c",
+    "@dimensiondev/holoflows-kit": "=0.8.0-20210429202414-74fb9a1",
     "@dimensiondev/kit": "=0.0.0-20210630045432-e05621c",
     "@masknet/icons": "workspace:*",
     "@masknet/plugin-example": "workspace:*",
