+++ packages/dashboard/src/initialization/Dashboard.tsx
@@ -4,13 +4,13 @@ import { StylesProvider } from '@material-ui/styles'
 import {
     CustomSnackbarProvider,
     MaskLightTheme,
-    ErrorBoundary,
     applyMaskColorVars,
     MaskDarkTheme,
     useSystemPreferencePalatte,
 } from '@masknet/theme'
 import { ChainId } from '@masknet/web3-shared'
 import { Emitter } from '@servie/events'
+import { ErrorBoundary } from '@masknet/shared'
 
 import i18n from 'i18next'
 import { I18nextProvider } from 'react-i18next'
