+++ packages/dashboard/src/initialization/i18n.ts
@@ -1,5 +1,5 @@
 import i18n from 'i18next'
-import { addMaskThemeI18N } from '@masknet/theme'
+import { addMaskSharedI18N } from '@masknet/shared'
 import { addDashboardI18N } from '../locales'
 import { initReactI18next } from 'react-i18next'
 
@@ -10,5 +10,5 @@ i18n.init({
     fallbackLng: 'en',
 })
 i18n.use(initReactI18next)
-addMaskThemeI18N(i18n)
+addMaskSharedI18N(i18n)
 addDashboardI18N(i18n)
