+++ packages/dashboard/src/pages/Wallets/components/AddTokenDialog/index.tsx
@@ -1,5 +1,6 @@
 import { memo, useState } from 'react'
-import { MaskDialog, useSnackbarCallback } from '@masknet/theme'
+import { MaskDialog } from '@masknet/theme'
+import { useSnackbarCallback } from '@masknet/shared'
 import { ERC20TokenDetailed, useERC20TokenBalance, useERC20TokenDetailed, useWallet } from '@masknet/web3-shared'
 import { useUpdateEffect } from 'react-use'
 import { PluginServices } from '../../../../API'
