+++ packages/maskbook/package.json
@@ -8,7 +8,7 @@
   "dependencies": {
     "@balancer-labs/sor": "^1.0.0",
     "@dimensiondev/common-protocols": "1.6.0-20201027083702-d0ae6e2",
-    "@dimensiondev/holoflows-kit": "0.8.0-20210317064617-6c4792c",
+    "@dimensiondev/holoflows-kit": "=0.8.0-20210429202414-74fb9a1",
     "@dimensiondev/kit": "=0.0.0-20210630045432-e05621c",
     "@dimensiondev/metamask-extension-provider": "3.0.6-20210519045409-1835d4d",
     "@dimensiondev/stego-js": "=0.11.1-20201027083223-8ab41be",
