+++ packages/maskbook/src/UIRoot.tsx
@@ -3,7 +3,7 @@ import { StylesProvider } from '@material-ui/styles'
 import { CssBaseline, StyledEngineProvider, ThemeProvider } from '@material-ui/core'
 import { Web3Provider } from '@masknet/web3-shared'
 import { SnackbarProvider } from '@masknet/theme'
-import { ErrorBoundary, ErrorBoundaryBuildInfoContext } from '@masknet/theme'
+import { ErrorBoundary, ErrorBoundaryBuildInfoContext } from '@masknet/shared'
 import i18nNextInstance from './utils/i18n-next'
 import { useClassicMaskTheme } from './utils/theme'
 import { Web3Context } from './web3/context'
