+++ packages/maskbook/src/extension/options-page/DashboardDialogs/Persona/Backup.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react'
 import { User } from 'react-feather'
 import { useI18N, compressBackupFile, encodeArrayBuffer, encodeText } from '../../../../utils'
-import { QRCode } from '@masknet/theme'
+import { QRCode } from '@masknet/shared'
 import Services from '../../../service'
 import AbstractTab, { AbstractTabProps } from '../../../../components/shared/AbstractTab'
 import ShowcaseBox from '../../DashboardComponents/ShowcaseBox'
