+++ packages/maskbook/src/extension/options-page/DashboardDialogs/Wallet/ReceiveTab.tsx
@@ -2,7 +2,7 @@ import { IconButton, InputAdornment, makeStyles, TextField, Theme } from '@mater
 import FileCopyOutlinedIcon from '@material-ui/icons/FileCopyOutlined'
 import { useCopyToClipboard } from 'react-use'
 import { useI18N } from '../../../../utils'
-import { QRCode } from '@masknet/theme'
+import { QRCode } from '@masknet/shared'
 import type { Wallet } from '@masknet/web3-shared'
 import { useSnackbarCallback } from '../Base'
