+++ packages/plugin-infra/package.json
@@ -6,7 +6,7 @@
   "types": "./dist/",
   "scripts": {},
   "dependencies": {
-    "@dimensiondev/holoflows-kit": "0.8.0-20210317064617-6c4792c",
+    "@dimensiondev/holoflows-kit": "=0.8.0-20210429202414-74fb9a1",
     "@masknet/shared": "workspace:*",
     "@masknet/theme": "workspace:*",
     "@masknet/web3-shared": "workspace:*",
