+++ packages/shared/src/UI/components/index.ts
@@ -1,2 +1,4 @@
 export * from './TokenIcon'
 export * from './TokenList'
+export * from './QRCode'
+export * from './ErrorBoundary'
