+++ packages/shared/src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useValueRef'
 export * from './useObservableMapSet'
 export * from './useMenu'
+export * from './useSnackbarCallback'
