+++ packages/shared/src/index.ts
@@ -14,3 +14,4 @@ export * from './wallet'
 export * from './Persona/type'
 export * from './Persona/type'
 export * from './UI'
+export * from './locales'
