+++ packages/shared/src/locales/index.ts
@@ -1,9 +1,10 @@
+import { createI18NBundle } from '../i18n/register-ns'
+
 export * from './i18n_generated'
-import { createI18NBundle } from '@masknet/shared'
 import en from './en.json'
 import zh from './zh.json'
 import ja from './ja.json'
-export const addMaskThemeI18N = createI18NBundle('theme', {
+export const addMaskSharedI18N = createI18NBundle('theme', {
     en,
     zh,
     ja,
