+++ packages/shared/src/locales/ja.json
