+++ packages/shared/src/locales/zh.json
