+++ packages/shared/tsconfig.json
@@ -6,9 +6,9 @@
     "experimentalDecorators": true,
     "tsBuildInfoFile": "./dist/.tsbuildinfo"
   },
-  "include": ["./src/**/*", "./src/**/*.json"],
+  "include": ["./src/**/*", "./src/**/*.json", "./src/locales/*.json"],
   "ts-node": {
     "compilerOptions": { "module": "CommonJS" }
   },
-  "references": [{ "path": "../web3-shared" }]
+  "references": [{ "path": "../web3-shared" }, { "path": "../theme" }]
 }
