+++ packages/theme/README.md
@@ -1,10 +1,3 @@
 # Mask Theme
 
 This is shared theme and components. Preview all components in this package at <https://compassionate-northcutt-326a3a.netlify.app/storybook-static/theme/>.
-
-To use this package, you must setup `i18next`.
-
-```ts
-import { addMaskThemeI18N } from '@masknet/theme'
-addMaskThemeI18N(your_i18next_instance)
-```
