+++ packages/theme/src/Components/index.ts
@@ -1,6 +1,4 @@
-export * from './ErrorBoundary/index'
 export * from './Dialogs/index'
 export * from './ButtonGroupTab/index'
-export * from './QRCode/index'
 export * from './Snackbar'
 export * from './SearchableList'
