+++ packages/theme/src/hooks/index.ts
@@ -1,2 +1,1 @@
 export * from './useTabs'
-export * from './useSnackbarCallback'
